Finally got tar to actually create an archive.
Seems like the problem was that I was making the archive in the scratch folder rather than in
the actually directory it needed to end up in. Cd'ing to the cleaning directory before the tar command fixed that.

Now I need to remove the ../tmp stuff from the archive name and hopefully everything will be good.